requirements.txt
# 📦 requirements.txt - VERSÃO HÍBRIDA OTIMIZADA

# ===== FRAMEWORK PRINCIPAL =====
Flask==3.0.3
Flask-CORS==5.0.0

# ===== BANCO DE DADOS =====
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
PyMySQL==1.1.1

# ===== AUTENTICAÇÃO E SEGURANÇA =====
PyJWT==2.9.0
bcrypt==4.2.0
cryptography==43.0.1

# ===== VALIDAÇÃO E UTILITÁRIOS =====
marshmallow==3.22.0
python-dotenv==1.0.1
Werkzeug==3.0.4

# ===== UPLOAD E ARQUIVOS =====
Pillow==10.4.0
python-magic==0.4.27

# ===== CACHE E PERFORMANCE (OPCIONAL) =====
redis==5.1.1

# ===== AWS S3 (OPCIONAL) =====
boto3==1.35.39
botocore==1.35.39

# ===== HTTP E REQUESTS =====
requests==2.32.3
urllib3==2.2.3

# ===== LOGS E MONITORAMENTO =====
structlog==24.4.0

# ===== DESENVOLVIMENTO =====
python-dateutil==2.9.0.post0

# ===== TESTING (OPCIONAL) =====
pytest==8.3.3
pytest-flask==1.3.0

# ===== SISTEMA BANCÁRIO (OPCIONAL) =====
# Estas dependências são opcionais para o sistema bancário
# xmltodict==0.13.0
# lxml==5.3.0